+++ test/libsolidity/semanticTests/types/convert_fixed_bytes_to_uint_same_type.sol
@@ -1,10 +1,10 @@
 contract Test {
-    function bytesToUint(bytes32 s) public returns (uint h) {
-        return uint(s);
+    function bytesToUint(bytes32 s) public returns (uint256 h) {
+        return uint256(s);
     }
 }
 // ====
 // compileViaYul: also
 // compileToEwasm: also
 // ----
-// bytesToUint(bytes32): "abc2" -> left(0x61626332)
+// bytesToUint(bytes32): 0x61626332 -> left(0x61626332)
