+++ test/libsolidity/semanticTests/types/convert_uint_to_fixed_bytes_same_size.sol
@@ -1,5 +1,5 @@
 contract Test {
-    function uintToBytes(uint h) public returns (bytes32 s) {
+    function uintToBytes(uint256 h) public returns (bytes32 s) {
         return bytes32(h);
     }
 }
@@ -7,4 +7,4 @@ contract Test {
 // compileViaYul: also
 // compileToEwasm: also
 // ----
-// uintToBytes(uint): left(0x616263) -> left(0x616263)
+// uintToBytes(uint256): left(0x616263) -> left(0x616263)
